Update spendingTiem model + add spending history

+++ ToDoList/ToDoList/Models/SpendingItem.swift
@@ -4,7 +4,7 @@ import Foundation
 struct SpendingItem: Codable, Identifiable {
     
     let id: String
-    let amount: Int
+    let amount: Double
     let descriptionText: String
     let spendingType: SpendingTypology
     let creationDate: TimeInterval
@@ -12,7 +12,7 @@ struct SpendingItem: Codable, Identifiable {
     let category: CategorySpending
     let status: StatusTypology
     
-    init(id: String, amount: Int, descriptionText: String, spendingType: SpendingTypology, creationDate: TimeInterval, dateTask: TimeInterval, category: CategorySpending, status: StatusTypology) {
+    init(id: String, amount: Double, descriptionText: String, spendingType: SpendingTypology, creationDate: TimeInterval, dateTask: TimeInterval, category: CategorySpending, status: StatusTypology) {
         self.id = id
         self.amount = amount
         self.descriptionText = descriptionText
